// src/app/models/referral/budget.model.ts
export class BudgetModel {
    id: number;
    subCategory: string = '';
    january: number;
    february: number;
    march: number;
    april: number;
    may: number;
    june: number;
    july: number;
    august: number;
    september: number;
    october: number;
    november: number;
    december: number;
    total: number;
    year: number;
}